1. Finished testing GraphML reader/writer - they now support accept/reject, highlight and colour. 2. Wrote the routine to stop incompatible states from being entered into a matrix (note that this does not prevent seemingly compatible vertices from receiving a score of -1 on merge, just makes this less likely by eliminating obviously incompatible candidates). 3. Matrix-based compatibility score computation now runs properly on a few tests.
git-svn-id: svn://svn.code.sf.net/p/statechum/code/XMachineTool/trunk@269 4b8d0e13-f52a-0410-bc0e-8c3b9955e25a

+++ statechum/analysis/learning/experiments/TestAbstractExperiment.java
@@ -72,7 +72,15 @@ public class TestAbstractExperiment {
 			for(File f:directory.listFiles())
 			{
 				Assert.assertFalse(f.isDirectory());
-				Assert.assertTrue("cannot delete file "+f,f.delete());
+				if (!f.delete())
+				{
+					try {
+						Thread.sleep(1000);// wait for a sec
+					} catch (InterruptedException e) {
+						// ignore
+					}
+					Assert.assertTrue("cannot delete file "+f,f.delete());
+				}
 			}
 			directory.delete();
 		}
